move arrow logic into new file
